Add files via upload
Amazon closing sales forecasting is a time series project which uses ARIMA to model the relationship  between Amazon closing stock prices  and historical Closing Stock price . The goal of the project is to build an understanding of how ARIMA model can be used for time series forecasting. The dataset is available on Kaggle :
https://www.kaggle.com/datasets/varpit94/amazon-stock-data/download?datasetVersionNumber=8
ARIMA model had an RMSE of 141. For a more accurate prediction other complex models  like RNN can be used